+++ apps/frontend/src/app/3_organisms/RedemptionsHistoryFrame/RedemptionsHistoryFrame.tsx
@@ -24,7 +24,11 @@ import { ExportCSV } from '../../2_molecules/ExportCSV/ExportCSV';
 import { useNotificationContext } from '../../../contexts/NotificationContext';
 import { useAccount } from '../../../hooks/useAccount';
 import { translations } from '../../../locales/i18n';
-import { Bitcoin, EXPORT_RECORD_LIMIT } from '../../../utils/constants';
+import {
+  Bitcoin,
+  DEFAULT_HISTORY_FRAME_PAGE_SIZE,
+  EXPORT_RECORD_LIMIT,
+} from '../../../utils/constants';
 import {
   Redemption,
   Redemption_Filter,
@@ -34,16 +38,13 @@ import { dateFormat } from '../../../utils/helpers';
 import { formatValue } from '../../../utils/math';
 import { useGetRedemptionsHistory } from './hooks/useGetRedemptionsHistory';
 
-// TODO usage example, to be removed
-const DEFAULT_PAGE_SIZE = 10;
-
 export const RedemptionsHistoryFrame: FC = () => {
   const { t } = useTranslation();
   const { account } = useAccount();
   const { addNotification } = useNotificationContext();
 
   const [page, setPage] = useState(0);
-  const [pageSize, setPageSize] = useState(DEFAULT_PAGE_SIZE);
+  const [pageSize, setPageSize] = useState(DEFAULT_HISTORY_FRAME_PAGE_SIZE);
   const chain = chains.find(chain => chain.id === defaultChainId);
 
   const [orderOptions, setOrderOptions] = useState<OrderOptions>({
@@ -251,7 +252,7 @@ export const RedemptionsHistoryFrame: FC = () => {
         getData={exportData}
         filename="transactions"
         className="mb-7 hidden lg:inline-flex"
-        onExportEnd={() => setPageSize(DEFAULT_PAGE_SIZE)}
+        onExportEnd={() => setPageSize(DEFAULT_HISTORY_FRAME_PAGE_SIZE)}
         disabled={!redemptions || redemptions.length === 0}
       />
       <div className="bg-gray-80 py-4 px-4 rounded">
