+++ apps/frontend/src/app/3_organisms/TransactionHistoryFrame/TransactionHistoryFrame.tsx
@@ -27,8 +27,9 @@ import { useAccount } from '../../../hooks/useAccount';
 import { translations } from '../../../locales/i18n';
 import {
   Bitcoin,
-  EXPORT_RECORD_LIMIT,
   LIQUIDATION_RESERVE_AMOUNT,
+  DEFAULT_HISTORY_FRAME_PAGE_SIZE,
+  EXPORT_RECORD_LIMIT,
 } from '../../../utils/constants';
 import {
   InputMaybe,
@@ -41,15 +42,12 @@ import { dateFormat } from '../../../utils/helpers';
 import { formatValue } from '../../../utils/math';
 import { useGetTroves } from './hooks/useGetTroves';
 
-// TODO usage example, to be removed
-const DEFAULT_PAGE_SIZE = 10;
-
 export const TransactionHistoryFrame: FC = () => {
   const { t } = useTranslation();
   const { account } = useAccount();
   const { addNotification } = useNotificationContext();
   const [page, setPage] = useState(0);
-  const [pageSize, setPageSize] = useState(DEFAULT_PAGE_SIZE);
+  const [pageSize, setPageSize] = useState(DEFAULT_HISTORY_FRAME_PAGE_SIZE);
   const chain = chains.find(chain => chain.id === defaultChainId);
   const [filters, setFilters] = useState<InputMaybe<TroveChange_Filter>>({});
 
@@ -591,7 +589,7 @@ export const TransactionHistoryFrame: FC = () => {
         getData={exportData}
         filename="transactions"
         className="mb-7 hidden lg:inline-flex"
-        onExportEnd={() => setPageSize(DEFAULT_PAGE_SIZE)}
+        onExportEnd={() => setPageSize(DEFAULT_HISTORY_FRAME_PAGE_SIZE)}
         disabled={!troves}
       />
       <div className="bg-gray-80 py-4 px-4 rounded">
