+++ apps/frontend/src/utils/constants.ts
@@ -60,6 +60,8 @@ export const servicesConfig = {
 };
 
 export const EXPORT_RECORD_LIMIT = 500;
+export const DEFAULT_HISTORY_FRAME_PAGE_SIZE = 10;
+
 export const DEFAULT_TIMEOUT_SECONDS = 5000;
 export const LIQUIDATION_RESERVE_AMOUNT = 20;
 export const GAS_LIMIT_OPEN_TROVE = 1_000_000;
